fcos: adapt vault-agent config
The terraform does not generate a new wrapped secretid with every run,
but rewrites the wrapped token env file, which was deleted previously as
a marker the token didn't had to be unwrapped again. Now the unwrap
service will not overwrite the secretid if unwrapping fails. Also
removes Before/After relation between units, vault-agent can handle the
secretid changing.